Validate build artifacts that originated remotely and can't be accessed locally, unless otherwise specified by the OutputService.
Previously, we assumed that if a file wasn't available locally and we know it was available remotely recently, the file is still accessible, but this isn't true when you've written a file to a remote OutputService and then switch to a local (or no) OutputService. This results in build failures when Bazel tries to read a file that it no longer can access that it should have validated during the analysis phase.

PiperOrigin-RevId: 302110160