In SequencedSkyframeExecutorTest#testThreeSharedActionsRacing, make sure action C has started before interrupting the main thread.
If the main thread really dallies, it might not enqueue C before the build is interrupted, causing a deadlock as action A waits for C's signal before finishing.

I verified in a debugger that without this fix, I could cause a deadlock like the one in the linked bug. However, the deadlock didn't repro for me in 2k runs without this fix, so I can't be totally positive that this fixes the issue.

PiperOrigin-RevId: 302211501