Provide error details if capabilities call fails
That the exception cause of the AbruptExitException is never printed anywhere,
so by default, when an error happens in this call, it usually looks like this:
```
ERROR: Failed to query remote execution capabilities: UNAVAILABLE: io exception
```

This is independent of whether the failure happened because the connection
failed or the ssl certificate did not validate. Needless to say, this is a
pretty bad user experience.

Change-Id: I846d1f5b6cd73f7663593b1292db3336d998ecfb