bazel packages: delete testCreatePackageIsolatedFromOuterErrors
This test, from CL 6198296 in 2008, ensured that an error injected
before the return of PackageFactory.parseBuildFile would not cause
PackageFactoryApparatus.createPackage to fail with an exception.

Although enthusiastically approved at the time by someone with the
same login name as me, this is not a style of test we want to
encourage, as it drills holes into lib.packages to support artificial
scenarios. It also requires a lot of support code.
Tests should be expressed in terms of stable interfaces.

Also:
- inline and delete PackageFactory.parseBuildFile.
  Enable option.recordScope in tests.
- delete ParsingTracker and ErrorReporter (part of above test,
  but artificially distanced from it by the awful Test/TestBase split).
- delete getEnvironmentExtensions() abstract method (always empty)
  and parameter of PackageFactoryApparatus
- delete assertOutputFileForRule method (dead code)
- delete getPathPrefix() (always empty)

RELNOTES: N/A
PiperOrigin-RevId: 303802864