Clean up nullability in the CommandEnvironment#pendingException field's value.
Distinguish between pendingException being unset, being set to a dummy value by #precompleteCommand, and being set to a not-dummy value BlazeModuleEnvironment#exit. Commit 54c9f1 inadvertently got rid of this distinction, since it disallowed the AbruptExitException ctor taking in a null ExitCode.

This distinction is potentially relevant for CommandEnvironment#getPendingExitCode. In this change here, I made that method private and cleaned up its usage.

The distinction was added pretty early in the history of Bazel in commit 6e5e8f. It didn't actually matter in the codebase back then, and it still doesn't now, because BuildTool#processRequest is called *before* CommandEnvironment#precompleteCommand (see the consecutive lines in BlazeCommandDispatcher...), but let's not make the code brittle in the face of future changes. I also wanted to save future code readers time; I spent ~10 mins convincing myself that the code was "correct" today.

RELNOTES: None
PiperOrigin-RevId: 303968383