Windows native launcher: support launching with 8dot3 Windows path
Windows native launcher finds the shell/python script to launch based on the first command line argument (the path name of itself). But it could be a shortened 8dot3 windows path. Eg. when `bazel test` runs a test, if the test binary path is too long (> 260), Bazel will convert it to a short path. The native launcher should convert it back to its full path before using it to find the script to launch.

Fixes https://github.com/bazelbuild/bazel/issues/11023

Closes #11047.

PiperOrigin-RevId: 303994833