Inject caller of AutoProfiler when using Flogger. Log sites look like:
200310 19:15:09.602:I 28 [com.google.devtools.build.lib.analysis.BuildView.update:221] Spent 4 milliseconds doing test description logged

Note that the log line is the last line of the "try" block, not the line the AutoProfiler object was created.

Also provide overloads that do not take a logger, and do not take a logger or a min time for logging, for the common case where the calling class's logger does not need to be used.

PiperOrigin-RevId: 304201047