Fix bazel zsh completion for users with b defined
If you have an `alias b` defined in your dotfiles, this will break the
bazel completion (see [0][] for an example). Basically the `b` will
expand to the contents of the alias before the function is parsed.

If you already have `b() ...` to define a function, and you already
have a function named `b`, this completion will override that function.

Therefore update the function name to follow the existing convention.

[0]: https://github.com/ohmyzsh/ohmyzsh/issues/6237