Ensure attribute name lengths are reasonable
Puts a sensible, but generous, bound on attribute name length in order to
prevent (most probably accidental) runaway name lengths. The expectation is
that no reasonable human is or should be typing out 128 char attribute
names. Instead anything getting to this length is likely machine generated,
and should be checked for accidental unbounded growth.

While this is technically a breaking change, it's expected that the allowance
is generous enough that it's a noop. Do note that the previous max was
MAX_INT or the JVM OOMing, whichever came first, which is something we clearly
shouldn't allow.

RELNOTES: A maximum attribute name length is 128 is enforced
PiperOrigin-RevId: 305086330