De-disable ErrorMessageTest#testErrorMessageWithUnreadableLogFile
Don't check the line separator in the exception stack trace - it's not under the control
of the class under test and it's platform dependent, which is why it was previously
failing on windows.

Whether or not the error message header should use platform-dependent line separators
is a separate question.

RELNOTES: None
PiperOrigin-RevId: 305205551