Do not emit "overwrote runfile" errors when the before/after artifact is the same.
This is a suggestion in the form of a CL.

These are causing a lot of spam in the console when running "blaze build" on some targets, and they seem to not actually be an issue.

This situation teaches people to ignore these errors, which then leads people to also ignore (or miss in the pile of errors) these errors for the case where previousStr!=artifactStr, that I imagine could actually cause issues.

PiperOrigin-RevId: 305549080