bazel syntax: avoid materializing Locations for iterator check
Mutable iterable values (such as lists) are temporarily rendered
shallowly immutable when one or more for-loops are actively iterating
over them. Prior to this change, the value's Mutability would maintain
a hash table mapping objects to a list of active for-loop locations.

This meant that every for-loop caused memory allocation, and required
a Location for the loop. In a future change, I plan to avoid
materializing Locations during parsing, which will allow us to improve
the fidelity of position information recorded in the syntax tree, and
also reduce the memory required by syntax trees. However, it does mean
that Locations will not be materialized except in case of error.

This change replaces the lists of locations in the hash table by a
counter. Also, it adds a method to the Freezable interface that
permits value classes to implement their own intrusive counter.
The two most important mutable iterable types, StarlarkList and Dict,
now do this.

The error message for mutation of an actively iterated value no
longer prints the location of the for-loop, but it is usually obvious
from the stack trace.

Also:
- delete MutabilityException; use a plain EvalException.
- EvalUtils.lock,unlock removed from public API.
- Mutability: encode isFrozen boolean using iterators==null.
- Mutability.checkMutable inlined. The inconsistency check is
  removed---it was recently unreachable by construction.
- MutabilityTest: delete tests of obsolete functions.
- Freezable.checkMutable() renamed to Starlark.checkMutable(Freezable).
  No need for it to be a mix-in any more.
- Report type of frozen mutable value in failed update exceptions.
PiperOrigin-RevId: 305732656