Expose ctx.attr.write Args content for analysistest when possible
Currently, analysistest can look at the contents of actions generated with ctx.actions.write if content is a string, but not if it's Args.

It still can't inspect the contents if the output contains TreeArtifacts, in which case that can't be computed at analysis time. In that case, this returns null, as was the previous behavior.

RELNOTES: Expose ctx.attr.write Args content for analysistest when possible
PiperOrigin-RevId: 306289304