Use ByteStrings instead of Strings for stdout/err
These are starting as opaque bytes in SynchronizedOutputStream (and/or whatever
comes before that) and finishing as bytes in a proto. No need to take an extra
trip through String which add encoding overhead and room for error.

Drive-by cleanup to delete `discardAll` because it was always false.

Also add a test for SynchronizedOutputStream which documents the minutia of the
current behavior.

RELNOTES: None
PiperOrigin-RevId: 306311512