Officially ban starlark transitions on --define by throwing a nice error instead of crashing (including suggestion for using build settings.)
This is a single instance of a larger set of issues with starlark transitions not gracefully handling complicated native option types (combinations of converts, allowMultiple, and types). Worth special casing this one because (1) it's by far the most obvious pain point (see #11117 and linked bug) and (2) it's a good place to advertise build settings.

PiperOrigin-RevId: 306435591