bazel skylarkbuildapi: replace BaseFunction with StarlarkCallable
BaseFunction is an abstract base class to simplify the implementation
of StarlarkCallable. It is overused and unnecessary in most cases.
StarlarkCallable is an interface that represents all callable Starlark
values. This change replaces BaseFunction throughout Bazel's Starlark API
with StarlarkCallable, allowing these places in the API that accept a
function to work equally well with Starlark or built-in functions:

  Args.add(map_each)
  Args.add(map_all)
  rule(implementation)
  aspect(implementation)
  transition(implementation)

The StarlarkCallable interface does not expose a FunctionSignature
(an inessential concept I would like to delete). Consequently, the
eager check done by Args.add to ensure that its map_each
value is a function of one argument can no longer be implemented.
This is no great loss: signature mismatch is only one of a million
way a function call can fail, and it is the easiest to detect
dynamically as it happens on every call. In general we should
strive to treat callable values opaquely.

Similarly, some places that used to accept only a StarlarkFunction
(that is, a function defined in Starlark) also now accept any
StarlarkCallable. The only places that continue to demand a specific
non-interface type are places like rule(outputs), which must
introspect on the parameter names of a function defined in Starlark.

BUG=153667498
PiperOrigin-RevId: 306452316