In repo rules, don't warn about generator_* attributes being non-canonical
A recent change added the implicit generator_* attributes ("generator_name",
"generator_function", and possibly "generator_location") to all repo rules.
This had the unfortunate side-effect of spamming all users of certain repo
rules with a warning that the values for these attributes are non-canonical.

Since these attributes are implicitly created and supplied by Bazel itself,
and since they do not (or should not) affect repo rule determinism, there is
no action needed on the part of the user. This change hardcodes them to be
excluded from the canonical warning.

Fixes #11040. Closes #11113.

PiperOrigin-RevId: 306478566