bazel analysis: add verbose assertion for diagnosis of latent NPE bug
Add a preconditions check for a null starlark option as a response to http://b/153053875. I couldn't successfully repro the failure so adding this check in here to note the latent bug.

The logic here is complicated by aliases - also tried to add some helpful comments along the way for details I have to relearn every time I look at this code.

PiperOrigin-RevId: 306705250