Automated rollback of commit 886d7d590fae2f860788a4c8ccb4624e59fc212d.
*** Reason for rollback ***

b/154243114

Fixes https://github.com/bazelbuild/bazel/issues/11125

*** Original change description ***

Use ByteStrings instead of Strings for stdout/err

These are starting as opaque bytes in SynchronizedOutputStream (and/or whatever
comes before that) and finishing as bytes in a proto. No need to take an extra
trip through String which add encoding overhead and room for error.

Drive-by cleanup to delete `discardAll` because it was always false.

Also add a test for SynchronizedOutputStream which documents the minutia of the
current behavior.

RELNOTES: None
PiperOrigin-RevId: 306913995