When we aren't inlining StarlarkImportLookupValue, inline ASTFileLookupValue.
In this situation (note: this is the situation for Bazel), a separate ASTFileLookupValue node is unnecessary. A code comment block explains why in more detail. This code comment block also explains why we want to keep the old behavior of *not* inlining ASTFileLookupValue when we are inlining StarlarkImportLookupValue (note: Bazel doesn't do this). Note that we still need to keep ASTFileLookupFunction around in some form no matter what anyway because of the implementation of Bazel's `tools/build_rules/prelude_bazel` feature (see PackageFunction's usage).

Benchmarking this change on some internal targets showed up to a 4% retained heap improvement, depending on how much bzl code the targets use.

RELNOTES: None
PiperOrigin-RevId: 307229360