Redirect android persistent worker stdout/stderr to buffer
In certain circumstances the various android compilation tools could output to stdout while operating as a persistent worker which would lead to a corruption of the response protobuf and ultimately lead to a failed build.

This PR solves that problem by redirecting all stdout and stderr emitted by the worker to an in-memory buffer that is later attached to the proto response as an output. In this way, all future tools added to the tool wrapper can benefit from having their output logged rather than cause a build failure.

https://github.com/bazelbuild/bazel/issues/10527
https://github.com/bazelbuild/bazel/issues/9207

Closes #10911.

PiperOrigin-RevId: 307252644