Undo accidentally incompatible change in treatment of '(' and ')' in some globs.
Historically, if the glob pattern was sufficiently complex to require compiling to regexp
(i.e. it wasn't of the form "*foo" or "foo*"), Bazel's undocumented (and wrong) behavior
was to treat the parentheses as no-ops (if the parentheses were paired) or crash (if unpaired).

Unfortunately, it turns out some BUILD files have accidentally relied on this behavior - e.g. calling `glob(["(*.txt)"])` and expecting to match "foo.txt" instead of "(foo.txt)".

We will therefore have to undo treating '(' and ')' as literal characters for now - while still avoiding the crash; changing the behavior would require a flag.

RELNOTES: None.
PiperOrigin-RevId: 307400998