bazel syntax: reduce stack usage in evaluator recursion
This change breaks each case in the doEval(Expression) switch
into a separate child function to reduce the stack required
by a recursive call to eval. The frame size of a switch is
that of its largest case, so deeply nested expressions such
as a+...+z consume much more stack than necessary for a mere
addition.

Empirically this causes stack overflow to occur at a recursion
depth of around 20K additions, up from around 3K before this change.

BUG=153764542
PiperOrigin-RevId: 307403954