Cope with fsevents dropping events.
It is perfectly normal for the fsevents API to coalesce or drop events,
and when that happens, we are told that we must rescan the directory
contents ourselves.

However, our usage of this API did not handle this condition, which
means we were indeed losing events and not knowing about it.  To fix
this, monitor for this condition and, when we encounter it, simply
assume that everything has changed and give up with the diff.

This should fix the flakiness observed in MacOSXFsEventsDiffAwarenessTest
but at this point I don't know if it will.  I haven't been able to
reproduce the flakiness "as is", but I was able to spot this new
problem by writing a new test that touches a lot of files in a very
short period of time.

Fixes #10776.

RELNOTES: None.
PiperOrigin-RevId: 307466746