Fix handling of license and distribs attributes
Rule.getLicense() and Rule.getAttribute("license") could fall out of sync if a
the package-default value was set via `licenses` or `package` _after_ creating
a rule. Same goes for distribs. This fixes that by special-casing the licenses
and distribs attributes in AggregatingAttributeMapper to get the values from
the Rule itself, the same way we do for visibility.

There are still some lingering bugs around licenses and distribs consistency -
things using Package.Builder to check license visibility can be incorrect,
and BuildOutputFormatter and XmlOutputFormatter still get the attribute source
wrong.

PiperOrigin-RevId: 307523307