Fix SkylarkConstructor documentation collection
This was broken due to refactoring work pertaining to @SkylarkCallable annotation (and not caught due to lack of test coverage). The original regression was caused by change of behavior of CallUtils.collectSkylarkMethodsWithAnnotation. Prior to the refactoring, that method would return the selfCall method of the requested module. After that refactoring, it would not. Selfcall methods were thus dropped from documentation.

Fixes #10848. Original regression was likely CL/277416070.

RELNOTES: None.
PiperOrigin-RevId: 307654334