Remove "low level" Package.Builder constructor
Trying to remove unnecessary indirection to make it more clear what can be made
immutable in packages.

Drive-by changes to use a function ref for the event handler and make the
duplicate label check more clear snuck in. The former got my attention since
it was right next to the constructors, the latter because it was touched by
the function ref threading.

PiperOrigin-RevId: 307694188