Open and read bzl files at most once per build.
Currently Bazel is opening and reading bzl files twice: once to get their contents, and once to get their digest. Instead we can simply grab the precomputed digest from the FileValue we already have, or we can compute it on the fly from the contents we just read. Therefore, this CL is a [small] strict performance win.

RELNOTES: None
PiperOrigin-RevId: 307813465