Remove misleading setting of applicable_licenses default attribute values
This was essentially dead code - builder.setDefaultApplicableLicenses was
setting applicableLicenses directly on the package instance, then the default
attribute population code was using the builder's defaultApplicableLicenses,
which was never set. Note that "fixing" this by setting the builders
defaultApplicableLicenses is unsound because the builder's default value
can change during the course of package loading, leading to inconsistent
defaults. Instead I opted to wipe the slate clean so we can start from scratch.

I also went ahead and deleted the method for special-casing default values and
reduced how far along we thread the package builder in order to avoid the
temptation of using it again.

PiperOrigin-RevId: 307837806