bazel syntax: delete nearly all of SkylarkType
Most of it became dead code since commit 858039031b3fcfb58ab589e06b1e6db493a9e740
removed the cast operators and commit 7f266408e95c50af158df1108fe6bffc3030eec7
stopped using it for validating arguments to
built-in function calls. The Union and Combination
classes were entirely dead, though it was hard to
see through all the overloaded methods.

This change simplifies it to a wrapper around
a class symbol representing the top-level constructor
of the elements of a depset, or null for an empty
depset.

PiperOrigin-RevId: 307858013