Semi-automated rollback of commit 3b3c89dd8de6fff27531a026f6a6faaeefd9cc6a.
*** Reason for rollback ***

Rolling back with fix for Kythe's usage.

Kythe doesn't call DigestHashFunction#setDefault. This may be true in general for other PackageLoader client (esp. in OSS-land with BazelPackageLoader).

As a workaround, we have AbstractPackageLoader get the DigestHashFunction from the FileSystem the client passed in. We then thread this through everywhere. For the regular Blaze/Bazel codebase, we can use DigestHashFunction#getDefaultUnchecked because GoogleUnixFileSystemModule and BazelFileSystemModule have already called DigestHashFunction#setDefault (lots of other pieces of the codebase assume this).

RELNOTES: None
PiperOrigin-RevId: 307859624