Allow ninja_build .d file processing to handle undeclared inputs
This introduces a new Artifact type, "NinjaMysteryArtifact", which represents a generated artifact that was not explicitly declared as any action output, though it is known to have been generated as a side effect due to other info in the graph (for example, post-execution .d files of other actions).

RELNOTES: None.
PiperOrigin-RevId: 307908968