bazel syntax: make FunctionSignature private
This change makes FunctionSignature private; for ease of review,
it will be inlined, simplified, and deleted in a follow-up CL.
FunctionSignature.fromParameters has moved to Parser;
its logic has not changed.

FunctionSignature's only external clients are ApiExporter and Skydoc,
and they only need four pieces of information: the parameter names,
hasVarargs, hasKwargs, and a getter for (optional) default values,
so that is all that StarlarkFunction exposes.

PiperOrigin-RevId: 308106753