bazel syntax: StarlarkFunction.toString: don't print default values
This change causes the toString method of StarlarkFunction not to
include the default values of optional parameters. The default values
are potentially large and they are usually not very interesting.
(I don't know why we print even the parameter names. Python doesn't.)

This string is not accessible to Starlark programs; it is only a
(breaking) API change. One Copybara test needed updating.

PiperOrigin-RevId: 308113924